Closing issue #67
Update was not occurring with power spectral density band reference in
the single view mode.  Possible that other working branch was not
pulled in when this was resolved the first time.  Update calls to
secondary display axes as well so that all 5 power bands will be
displayed when selected and also improved secondary axes display to
remove vector magnitude axis when switching from psd feature and
others.