updates for issue #87
part 2: the main figure should recognize that a new batch output folder
exists or an older when has been updated, and ask the user if the
recent data should be used in place of any data currently loaded.
-
Solution: Check occurs now when switching to results view and a stat
tool instance already exists.  If the instance’s results path differs
from the controller’s source path the user is prompted if they want to
update the results view the controller’s path which would have been
updated after a call to batch tool.

- Another solution would be to have existing PAStattool’s instance’s
listen for batch tool completions, but this would also require updates
to the controller code.